refactor room amenity manager
