+++ app/src/androidTest/java/com/xepicgamerzx/hotelier/HotelAmenityManagerTest.java
@@ -14,7 +14,7 @@ import com.xepicgamerzx.hotelier.objects.hotel_objects.Hotel;
 import com.xepicgamerzx.hotelier.objects.hotel_objects.HotelRoom;
 import com.xepicgamerzx.hotelier.objects.hotel_objects.HotelAmenitiesEnum;
 import com.xepicgamerzx.hotelier.storage.hotel_reference_managers.HotelAmenitiesCrossManager;
-import com.xepicgamerzx.hotelier.storage.hotel_reference_managers.HotelAmenityManager;
+import com.xepicgamerzx.hotelier.storage.hotel_managers.HotelAmenityManager;
 import com.xepicgamerzx.hotelier.storage.hotel_managers.HotelManager;
 import com.xepicgamerzx.hotelier.storage.hotelier_database.HotelierDatabase;
 import com.xepicgamerzx.hotelier.storage.hotel_managers.RoomManager;
