+++ app/src/androidTest/java/com/xepicgamerzx/hotelier/RoomAmenityManagerTest.java
@@ -15,7 +15,7 @@ import com.xepicgamerzx.hotelier.objects.hotel_objects.RoomAmenity;
 import com.xepicgamerzx.hotelier.storage.hotel_managers.HotelManager;
 import com.xepicgamerzx.hotelier.storage.hotelier_database.HotelierDatabase;
 import com.xepicgamerzx.hotelier.storage.hotel_reference_managers.RoomAmenitiesCrossManager;
-import com.xepicgamerzx.hotelier.storage.hotel_reference_managers.RoomAmenityManager;
+import com.xepicgamerzx.hotelier.storage.hotel_managers.RoomAmenityManager;
 import com.xepicgamerzx.hotelier.storage.hotel_managers.RoomManager;
 
 import org.junit.After;
