+++ app/src/main/java/com/xepicgamerzx/hotelier/storage/hotel_reference_managers/HotelAmenitiesCrossManager.java
@@ -5,6 +5,7 @@ import android.app.Application;
 import com.xepicgamerzx.hotelier.objects.hotel_objects.Hotel;
 import com.xepicgamerzx.hotelier.objects.cross_reference_objects.HotelAmenitiesCrossRef;
 import com.xepicgamerzx.hotelier.objects.hotel_objects.HotelAmenity;
+import com.xepicgamerzx.hotelier.storage.hotel_managers.HotelAmenityManager;
 import com.xepicgamerzx.hotelier.storage.hotelier_database.HotelierDatabase;
 import com.xepicgamerzx.hotelier.storage.dao.HotelAmenitiesCrossDao;
 import com.xepicgamerzx.hotelier.storage.hotel_managers.HotelManager;
