+++ app/src/main/java/com/xepicgamerzx/hotelier/storage/hotel_reference_managers/RoomAmenitiesCrossManager.java
@@ -6,6 +6,7 @@ import com.xepicgamerzx.hotelier.objects.hotel_objects.HotelRoom;
 import com.xepicgamerzx.hotelier.objects.cross_reference_objects.RoomAmenitiesCrossRef;
 import com.xepicgamerzx.hotelier.objects.hotel_objects.RoomAmenity;
 import com.xepicgamerzx.hotelier.storage.dao.RoomAmenitiesCrossDao;
+import com.xepicgamerzx.hotelier.storage.hotel_managers.RoomAmenityManager;
 import com.xepicgamerzx.hotelier.storage.hotel_managers.RoomManager;
 import com.xepicgamerzx.hotelier.storage.hotelier_database.HotelierDatabase;
